Longer integer for chunk math. Int32 overflow when calculating file diff

+++ DepotDownloader/ContentDownloader.cs
@@ -792,7 +792,7 @@ namespace DepotDownloader
                         }
                         else
                         {
-                            size_downloaded += (file.TotalSize - (ulong)neededChunks.Select(x => (int)x.UncompressedLength).Sum());
+                            size_downloaded += (file.TotalSize - (ulong)neededChunks.Select(x => (long)x.UncompressedLength).Sum());
                         }
                     }
